rename auth store
