+++ src/app/admin/pages/dashboard/dashboard.component.html
@@ -5,3 +5,8 @@
 <!-- <router-outlet></router-outlet> -->
 <h1>Dashboard</h1>
 <h3>Hi {{ currentUser?.first_name }}</h3>
+<!-- <h3>{{ currentUser?.user_profile[0].bio }}</h3> -->
+
+<!-- <div class="col-sm-12" *ngIf="(currentUser?.user_profile[0] | keyvalue)?.length">
+  <h3>{{ currentUser?.user_profile[0].bio }}</h3>
+</div> -->
