+++ src/app/auth/state/auth.action.ts
@@ -2,7 +2,7 @@ import { User } from "../../core/models/user";
 
 /* NgRx */
 import { Action } from "@ngrx/store";
-import { UserEffects } from "./user.effects";
+import { UserEffects } from "./auth.effects";
 
 export enum UserActionTypes {
   LOGIN = "[User] Login",
