+++ src/app/shared/layout/header/header.component.ts
@@ -2,9 +2,9 @@ import { Component, OnInit } from "@angular/core";
 import { AuthService } from "../../../core/services/user.service";
 import { Router } from "@angular/router";
 import { User } from "../../../core/models/user";
-import * as userActions from "../../../auth/state/user.action";
+import * as userActions from "../../../auth/state/auth.action";
 import { Store } from "@ngrx/store";
-import * as fromUser from "../../../auth/state/user.reducer";
+import * as fromUser from "../../../auth/state/auth.reducer";
 
 @Component({
   selector: "app-header",
